Fallback when unable to initialize the clipboard
Currently vizia will crash if it is unable to create a ClipboardProvider
using the default provider for the platform. On Linux that is X11, so if
the display is using Wayland then the application will crash on startup.
This patch falls back to `NopClipboardContext` which does nothing.
The clipboard will not work but the application will not crash.

Fully implementing Wayland support will take more work as we need need a
`RawWidnowHandle` to request a proxy, which is not available until after
the `Context` is created.