Initial attempt at async-list
Adds an example containing a scrollable list which displays a virtual sub-view over the full list. The list items themselves trigger a computation asynchronously (increasing in time with list index), displaying different views depending on whether the computation is done. TODO: figure out a way to delay rendering so that fast computations which could be done synchronously don't cause flickering.